+++ collection/constraint.go
@@ -1,0 +1,6 @@
+package collection
+
+// Ordered is a constraint for types that can be ordered.
+type Ordered[E any] interface {
+	Compare(E) int
+}