+++ collection/constraint_test.go
@@ -1,0 +1,9 @@
+package collection_test
+
+import "cmp"
+
+type elem int
+
+func (e elem) Compare(other elem) int {
+	return cmp.Compare(e, other)
+}