+++ collection/doc.go
@@ -1,0 +1,5 @@
+// Package collection provides generic data structures.
+//
+// These structures are not intended as complete implementations; they only
+// provide the operations required by official Dogmatiq engines and utilities.
+package collection